refactor: split UI behaviours and noise gradient out of main.js
Extract the navbar, command palette, side dots, mobile menu, counters,
tagline reveal, and research carousel into js/ui.js, and the hero noise
shader into js/noise-gradient.js. Move the shared escapeHtml helper to
utils.js so ui.js and the main.js renderers share it without a circular
import. main.js drops from ~1670 to ~810 lines and now focuses on content
rendering plus DOMContentLoaded orchestration; the extracted modules are
re-exported so the test surface is unchanged.
